requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
httpx==0.27.0
python-multipart==0.0.9
psycopg2-binary==2.9.9
pydantic>=2.7
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-api-python-client==2.131.0